requirements.txt
mango-explorer
